requirements.txt
streamlit>=1.38,<2
yt-dlp>=2024.8.6
requests>=2.32
sqlalchemy>=2.0
openai-whisper>=20231117
